LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-position-multicol-fixed-001-expected.txt


PASS .target 1
PASS .target 2
FAIL .target 3 assert_equals:
<div class="target" data-offset-x="20" data-offset-y="10" data-expected-width="130" data-expected-height="50"></div>
width expected 130 but got 20
FAIL .target 4 assert_equals:
<div class="target" data-offset-x="20" data-offset-y="10" data-expected-width="130" data-expected-height="50"></div>
width expected 130 but got 20
